Extract currency from settings
