+++ README.md
@@ -60,21 +60,21 @@ After all, the config file at `config/money.php` should be modified for your own
     - [📄 Cloning](/docs/01_usage/cloning.md)
     - [🎯 Casting](/docs/01_usage/casting.md)
     - [🚨 Validation rule](/docs/01_usage/validation_rule.md)
-4. [⚙ Settings](/docs/02_settings/README.md)
+3. [⚙ Settings](/docs/02_settings/README.md)
     - [Decimals](/docs/02_settings/decimals.md)
     - [Thousands separator](/docs/02_settings/thousands_separator.md)
     - [Decimal separator](/docs/02_settings/decimal_separator.md)
     - [Ends with zero](/docs/02_settings/ends_with_zero.md)
     - [Space between currency and amount](/docs/02_settings/space_between.md)
-    - [Currency](/docs/02_settings/currency.md)
-5. [💲 Currencies](/docs/03_currencies/README.md)
+4. [💲 Currencies](/docs/03_currencies/README.md)
     - [Information](/docs/03_currencies/information.md)
     - [Position](/docs/03_currencies/position.md)
     - [Display](/docs/03_currencies/display.md)
     - [Currency List](/docs/03_currencies/currency_list.md)
     - [Preferred symbol](/docs/03_currencies/preferred_symbol.md)
     - [Get current currencies](/docs/03_currencies/get_currencies.md)
-6. [💵 Money](/docs/04_money/README.md)
+5. [💵 Money](/docs/04_money/README.md)
+    - [💲 Currency](/docs/04_money/currency.md)
     - Static methods
         - [Getters](/docs/04_money/static/getters.md)
         - [`Money::set()`](/docs/04_money/static/set.md)
@@ -85,7 +85,6 @@ After all, the config file at `config/money.php` should be modified for your own
         - Getters
             - [`getAmount()`](/docs/04_money/object/getAmount.md)
             - [`getPureAmount()`](/docs/04_money/object/getPureAmount.md)
-            - [`getCurrency()`](/docs/04_money/object/getCurrency.md)
         - Calculations
             - [`add()`](/docs/04_money/object/add.md)
             - [`subtract()`](/docs/04_money/object/subtract.md)
