+++ docs/02_settings/README.md
@@ -27,7 +27,6 @@ All the settings that are not provided or not changed will have default values, 
 3. [Decimal separator](/docs/02_settings/decimal_separator.md)
 4. [Ends with zero](/docs/02_settings/ends_with_zero.md)
 5. [Space between currency and amount](/docs/02_settings/space_between.md)
-6. [Currency](/docs/02_settings/currency.md)
 
 ---
