+++ docs/04_money/README.md
@@ -1,19 +1,19 @@
 # 💵 Money
-Here we are, prepared and ready to create our own Money objects.
+Here we are, prepared and ready to create our own monetary objects.
 
 There are *static* methods as well as *object* ones.
 
-1. Static methods
+1. [💲 Currency](/docs/04_money/currency.md)
+2. Static methods
     - [Getters](/docs/04_money/static/getters.md)
     - [`Money::set()`](/docs/04_money/static/set.md)
     - [`Money::of()`](/docs/04_money/static/of.md)
     - [`Money::correctInput()`](/docs/04_money/static/correctInput.md)
     - [`Money::parse()`](/docs/04_money/static/parse.md)
-2. Object methods
+3. Object methods
     - Getters
         - [`getAmount()`](/docs/04_money/object/getAmount.md)
         - [`getPureAmount()`](/docs/04_money/object/getPureAmount.md)
-        - [`getCurrency()`](/docs/04_money/object/getCurrency.md)
     - Calculations
         - [`add()`](/docs/04_money/object/add.md)
         - [`subtract()`](/docs/04_money/object/subtract.md)
