+++ docs/upgrade/3.x_to_4.x.md
@@ -182,7 +182,7 @@ All money construct methods:
 
 now accept only `string` as their first argument.
 
-## Methods
+## Money
 
 ### 🟧 `multiple()`
 
@@ -201,6 +201,16 @@ Now, with `strict` flag it compares not only amounts, but also currencies of mon
 
 The method was renamed to `Money::of()`, the rest is the same.
 
-### ❌ `bind()`, `unbind()`, `bound()` in `MoneySettings`
+### 🟩 `setCurrency()`
+
+The method was extracted from settings and moved here.
+
+## MoneySettings
+
+### ❌ `bind()`, `unbind()`, `bound()`
 
 These methods were removed completely because of [Garbage Collection](https://github.com/PostScripton/laravel-money/issues/45).
+
+### ❌ `getCurrency()`, `setCurrency()`
+
+These methods were extracted directly from settings into Money.
